Beings now steal from other houses
Unresolved issues when their house becomes stuffed with furniture and
have to search the whole map to find that there is nowhere to put
anything. Currently a bit of lag while they figure that out, and then
the code for picking a new house can lead to them first depositing their
currently held furniture in a nearby house, looking for a new piece not
inside their first house or at current location, and discovering anew
that there is nowhere to put anything. Also, two beings can get stuck in
a loop of sorts stealing from each others houses. Both of these issues
are potentially not an issue in "proper" play